+++ client/components/core/index.js
@@ -1,9 +1,11 @@
 import ApplicationRouter from './ApplicationRouter';
+import Footer from './Footer';
 import Head from './Head';
 import * as Misc from './misc';
 
 export {
     ApplicationRouter,
+    Footer,
     Head,
     Misc
 };