+++ client/components/index.js
@@ -1,5 +1,7 @@
 import * as Core from './core';
+import * as Pages from './pages';
 
 export {
-    Core
+    Core,
+    Pages
 };