+++ client/components/pages/home/sections/index.js
@@ -1,0 +1,11 @@
+import About from './About';
+import Experience from './Experience';
+import Landing from './Landing';
+import Service from './Service';
+
+export {
+    About,
+    Experience,
+    Landing,
+    Service
+};